Refactor `ProgressPlugin`
