+++ src/felupe/mechanics/plugins/__init__.py
@@ -1,7 +1,5 @@
 from ._characteristic_curve import CharacteristicCurvePlugin
-from ._progress import ProgressPlugin
 
 __all__ = [
     "CharacteristicCurvePlugin",
-    "ProgressPlugin",
 ]
